Author ES6 modules, use Rollup for bundling
Using both default exports and named exports is allowed by the
spec, but not recommended. If we have both, then the usage
with require() needs to access .default, which is breaking, and
also can mess with bundling/scope hoisting. I moved the exports to
named only, and the tests run the same as before.

Rollup config is handled by microbundle, which takes the options
directly from package.json fields.

Tests now use the files in dist/, since they are CommonJS and node
does not yet understand ES modules.

One annoying thing atm is that npm install adds many versions of
acorn, which leads to errors with microbundle. Using yarn fixes
that. There are no dependencies, so this shouldn't change anything
but it is something to consider.

In addition to the modules field, I added scripts for publishing
to npm and bumping the tag, plus the "files" option. It should
match the instructions at the end of the README, but we can omit
it.